Misc cleanup and exposed idleTimeout as a property, as well as matched the type with the size (it was uint16_t but was placed into 4 bytes after we adjusted the firmware)

+++ RileyLinkKit/PumpOpsSynchronous.swift
@@ -46,7 +46,7 @@ class PumpOpsSynchronous {
 
     public static let PacketKey = "com.rileylink.RileyLinkKit.PumpOpsSynchronousPacketKey"
 
-    private static let standardPumpResponseWindow: UInt16 = 180
+    private static let standardPumpResponseWindow: UInt32 = 180
     private let expectedMaxBLELatencyMS = 1500
     
     // After
@@ -64,7 +64,7 @@ class PumpOpsSynchronous {
         return PumpMessage(packetType: .carelink, address: pump.pumpID, messageType: messageType, messageBody: body)
     }
 
-    private func sendAndListen(_ msg: PumpMessage, timeoutMS: UInt16 = standardPumpResponseWindow, repeatCount: UInt8 = 0, msBetweenPackets: UInt8 = 0, retryCount: UInt8 = 3) throws -> PumpMessage {
+    private func sendAndListen(_ msg: PumpMessage, timeoutMS: UInt32 = standardPumpResponseWindow, repeatCount: UInt8 = 0, msBetweenPackets: UInt8 = 0, retryCount: UInt8 = 3) throws -> PumpMessage {
         let cmd = SendAndListenCmd()
         cmd.packet = RFPacket(data: msg.txData)
         cmd.timeoutMS = timeoutMS
@@ -328,7 +328,7 @@ class PumpOpsSynchronous {
     }
 
     internal func changeWatchdogMarriageProfile(_ watchdogID: Data) throws {
-        let commandTimeoutMS: UInt16 = 30_000
+        let commandTimeoutMS: UInt32 = 30_000
 
         // Wait for the pump to start polling
         let listenForFindMessageCmd = GetPacketCmd()
